requirements.txt
paramiko
pysocks
pycrypto